+++ rockset/__init__.py
@@ -29,3 +29,4 @@ from rockset.exceptions import ApiException
 from rockset.rockset_client import DevRegions, Regions, RocksetClient
 from rockset.document import Document
 from rockset.query_builder import Q, F, P, ParamDict
+from rockset.query_paginator import QueryPaginator